backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.11.0
Pillow==9.5.0
pipenv==2023.3.20
platformdirs==3.2.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
virtualenv==20.21.0
virtualenv-clone==0.5.7
